// src/API.ts
export type CreateTodoInput = {
  id?: string | null,
  description: string,
};
// ...
export type Todo = {
  __typename: "Todo",
  id?: string,
  description?: string,
  createdAt?: string,
  updatedAt?: string,
};